+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-dbExtension/legend-engine-xt-relationalStore-databricks/legend-engine-xt-relationalStore-databricks-grammar/src/test/java/org/finos/legend/engine/language/pure/grammar/test/TestDatabricksConnectionGrammarRoundtrip.java
@@ -19,8 +19,24 @@ import org.junit.Test;
 public class TestDatabricksConnectionGrammarRoundtrip extends TestGrammarRoundtrip.TestGrammarRoundtripTestSuite
 {
     @Test
-    public void testConnectionGrammar()
+    public void testDeltaLakeDatabaseConnection()
     {
-        test("###Connection\n");
+        test("###Connection\n" +
+                "RelationalDatabaseConnection simple::DatabricksConnection\n" +
+                "{\n" +
+                "  store: apps::pure::studio::relational::tests::dbInc;\n" +
+                "  type: Databricks;\n" +
+                "  specification: Databricks\n" +
+                "  {\n" +
+                "    hostname: 'hostname';\n" +
+                "    port: 'port';\n" +
+                "    protocol: 'protocol';\n" +
+                "    httpPath: 'httpPath';\n" +
+                "  };\n" +
+                "  auth: ApiToken\n" +
+                "  {\n" +
+                "    apiToken: 'token';\n" +
+                "  };\n" +
+                "}\n");
     }
 }
