+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-execution/legend-engine-xt-relationalStore-executionPlan-connection-tests/pom.xml
@@ -278,6 +278,12 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-xt-relationalStore-databricks-protocol</artifactId>
+            <version>4.21.1-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
 
         <!--
         <dependency>
