+++ legend-engine-xts-relationalStore/legend-engine-xt-relationalStore-generation/legend-engine-xt-relationalStore-grammar/src/test/java/org/finos/legend/engine/language/pure/grammar/test/TestRelationalConnectionGrammarRoundtrip.java
@@ -87,27 +87,6 @@ public class TestRelationalConnectionGrammarRoundtrip extends TestGrammarRoundtr
                 "}\n");
     }
 
-    @Test
-    public void testDeltaLakeDatabaseConnection()
-    {
-        test("###Connection\n" +
-                "RelationalDatabaseConnection simple::DatabricksConnection\n" +
-                "{\n" +
-                "  store: apps::pure::studio::relational::tests::dbInc;\n" +
-                "  type: Databricks;\n" +
-                "  specification: Databricks\n" +
-                "  {\n" +
-                "    hostname: 'hostname';\n" +
-                "    port: 'port';\n" +
-                "    protocol: 'protocol';\n" +
-                "    httpPath: 'httpPath';\n" +
-                "  };\n" +
-                "  auth: ApiToken\n" +
-                "  {\n" +
-                "    apiToken: 'token';\n" +
-                "  };\n" +
-                "}\n");
-    }
 
     @Test
     public void testDataSourceSpecConfigurations()
